+++ src/blocks/tickets-button/view.ts
@@ -1,4 +1,0 @@
-import { GET_TICKETS_STORE } from '@shared/consts';
-import { store } from '@wordpress/interactivity';
-
-store( GET_TICKETS_STORE );