+++ src/blocks/tickets-button/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog: Tickets Button
 
+## v2.0.0 - [June 3, 2026]
+
+-   Breaking: Convert block from interactive block to ACF block
+
 ## v1.0.0
 
 -   Init!